put drawnametag function in a seperate file
